Refactor: 비슷한 도서 프로필 생성 역할의 클래스 병합
